% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSampleFeatures.R
\name{addsampleFeatures}
\alias{addsampleFeatures}
\title{addsampleFeatures}
\usage{
addsampleFeatures(object, sample.data = NULL, id.col = "sample")
}
\arguments{
\item{object}{CNQuant or SigQuant class object}

\item{sample.data}{data.frame containing sample-level variables}

\item{id.col}{column containing sample identifiers (default: "sample")}
}
\value{
CNQuant or SigQuant object with updated samplefeatData
}
\description{
\code{addsampleFeatures} adds custom sample-level data to the samplefeatData field
of a \code{CNQuant} or \code{SigQuant} class object. This can be additional sample
information (e.g. purity, tumour type, etc.) that can be stored within the
object and used in downstream analysis.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  data(test.sample.features)
  cnobj <- createCNQuant(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- addsampleFeatures(cnobj,sample.data=test.sample.features)
  getSamplefeatures(cnobj)
}
