% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCosineSim.R
\name{calculateCosineSim}
\alias{calculateCosineSim}
\title{calculateCosineSim}
\usage{
calculateCosineSim(
  x = NULL,
  y = NULL,
  perm = FALSE,
  nperms = 1000,
  alternative = "two.sided"
)
}
\arguments{
\item{x}{A matrix or vector (must be matrix if y is NULL)}

\item{y}{Optional matrix or vector (default: NULL)}

\item{perm}{Logical to enable permutation testing (default: FALSE)}

\item{nperms}{Number of permutations to compute (default: 1000)}

\item{alternative}{tailed test type (default: "two.sided"). Additionally can
run either "greater" or "less" for corresponding one-sided tests}
}
\value{
Either a vector, matrix, or list depending on input and 'perm' option
}
\description{
Function to compute cosine similarity between various input types and
additionally perform permutation testing for cosine similarity significance
testing.
}
