% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calculateFeatures.R
\docType{methods}
\name{calculateFeatures}
\alias{calculateFeatures}
\alias{calculateFeatures,CNQuant-method}
\title{calculateFeatures}
\usage{
calculateFeatures(object, method = "drews", smooth.diploid = TRUE, cores = 1)

\S4method{calculateFeatures}{CNQuant}(object, method = NULL, smooth.diploid = TRUE, cores = 1)
}
\arguments{
\item{object}{CNQuant object}

\item{method}{Method to extract copy number features. Default is "drews".}

\item{smooth.diploid}{Binary variable indicating whether segments close to 2 should be collapsed to 2 and merged together. Default is TRUE.}

\item{cores}{Number of CPU threads/cores to utilise via doParallel. Default is 1.}
}
\value{
A CNQuant class object with extracted features stored in the "featData" slot
}
\description{
Extracts and returns copy number features from copy number profiles in a CNQuant object.
}
