% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/calculateSampleByComponentMatrix.R
\docType{methods}
\name{calculateSampleByComponentMatrix}
\alias{calculateSampleByComponentMatrix}
\alias{calculateSampleByComponentMatrix,CNQuant-method}
\title{calculateSampleByComponentMatrix}
\usage{
calculateSampleByComponentMatrix(object, method = "drews")

\S4method{calculateSampleByComponentMatrix}{CNQuant}(object, method = NULL)
}
\arguments{
\item{object}{CNQuant object}

\item{method}{Determines the mixture components used to calculate sum-of-posterior probabilities. Default is "drews".}
}
\value{
A CNQuant class object with sum-of-posterior probabilities stored in the "featFitting" slot
}
\description{
Calculates and returns a sample-by-component matrix from copy number features in a CNQuant object.
}
