% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancerSpecificSignatures}
\alias{cancerSpecificSignatures}
\title{cancerSpecificSignatures}
\format{
A length 20 named list of character vectors
}
\usage{
data(cancerSpecificSignatures)
}
\description{
List of 20 named character vectors describing the signatures associated with each
cancer type (as described in Drews et al. 2022 signature discovery). Vectors include
10 pan-cancer signatures plus any additional signatures specific to the selected cancer type.
}
\details{
Cancer types are limited to "BRCA", "BLCA", "OV", "LUAD", "GBM", "STAD", "SKCM", "UCEC", "HNSC",
"CESC", "LUSC", "SARC", "COAD", "KIRC", "LGG", "READ", "PRAD", "ESCA", "LIHC", & "TGCT"
}
\keyword{datasets}
