% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseArdNMF.R
\name{chooseArdNMF}
\alias{chooseArdNMF}
\title{chooseArdNMF}
\usage{
chooseArdNMF(
  runs = NULL,
  method = "modal",
  k = NULL,
  decision = "bdiv",
  specificRun = NULL
)
}
\arguments{
\item{runs}{An object of class PLACEHOLDER-CLASS containing ARD-NMF runs (typically generated by \code{\link[=calculateArdNMF]{calculateArdNMF()}})}

\item{method}{Method to select the optimal number of decomposed signatures (k) (Default: "modal")}

\item{k}{Manually select the number of signatures (k) to return from list of ARD-NMF solutions (Default: NULL)}

\item{decision}{Optimisation metric to select best solution from selected subset of K signatures (Default: "bdiv")}

\item{specificRun}{A character string specifying a specific run to select from list of ARD-NMF solutions}
}
\value{
A list containing signature definitions and supporting ARD-NMF results for optimal solution
}
\description{
Select a optimal ARD-NMF run as final result to return for de novo signature definition assignment
}
