% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINSignatureQuantification.R
\docType{package}
\name{CINSignatureQuantification-package}
\alias{CINSignatureQuantification}
\alias{CINSignatureQuantification-package}
\title{CINSignatureQuantification}
\description{
Allowing the simple and quick quantification of copy number signatures in
cancer samples from copy number profiles. The signatures are a readout of mutational
processes resulting in chromosomal instability (CIN). The methods from
Drews et al. (Nature, 2022) and Macintyre et al. (Nature Genetics, 2018) are included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markowetzlab/CINSignatureQuantification}
  \item Report bugs at \url{https://github.com/markowetzlab/CINSignatureQuantification/issues}
}

}
\author{
\strong{Maintainer}: Philip S Smith \email{philip.smith@cruk.cam.ac.uk} (\href{https://orcid.org/0000-0002-9306-1747}{ORCID})

Other contributors:
\itemize{
  \item Ruben M Drews \email{Ruben.Drews@cruk.cam.ac.uk} (\href{https://orcid.org/0000-0001-7360-4970}{ORCID}) [contributor]
  \item University of Cambridge \email{Florian.Markowetz@cruk.cam.ac.uk} (\href{https://orcid.org/0000-0002-2784-5308}{ORCID}) [copyright holder, funder]
}

}
