% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINSignatureQuantification.R
\docType{package}
\name{CINSignatureQuantification}
\alias{CINSignatureQuantification}
\title{CINSignatureQuantification}
\description{
Allowing the simple and quick quantification of copy number signatures in
cancer samples from copy number profiles. The signatures are a readout of mutational
processes resulting in chromosomal instability (CIN). It is thought as a one-stop
solution, combining multiple published solutions. At the moment the methods from
Drews et al. (Nature, 2022) and Macintyre et al. (Nature Genetics, 2018) are included.
}
