% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/clinPredictionDenovo.R
\docType{methods}
\name{clinPredictionDenovo}
\alias{clinPredictionDenovo}
\alias{clinPredictionDenovo,SigQuant-method}
\title{clinPredictionDenovo}
\usage{
clinPredictionDenovo(object, sampTrain, sigsTrain)

\S4method{clinPredictionDenovo}{SigQuant}(object, sampTrain, sigsTrain)
}
\arguments{
\item{object}{SigQuant object}

\item{sampTrain}{Vector of sample names that should be used for training the
classifier.}

\item{sigsTrain}{Vector with two signature names on which the prediction
should be based upon.}
}
\value{
A vector with "Signature <1> higher" or "Signature <2> higher" for
all samples in the input object.
}
\description{
The function takes signature activities based on Drews et al. methodology and
predicts patient's response based on user-specified pair of signatures. The
user should supply a vector of samples for training purposes. The function
then trains the classifier on these samples before applying it to all samples
and return the labels.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- createCNQuant(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- calculateFeatures(cnobj[1:10],method="drews")
  cnobj <- calculateSampleByComponentMatrix(cnobj)
  cnobj <- calculateActivity(cnobj)
  # Not a real predictor
  cnobj <- clinPredictionDenovo(cnobj,
             sampTrain = getSamples(cnobj[1:5]),
             sigsTrain = c("CX1","CX6"))
}
