% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CNobj.drews}
\alias{CNobj.drews}
\title{CNobj.drews}
\format{
A CNQuant class object using the "drews" methodology
}
\usage{
data(CNobj.drews)
}
\description{
Example CNQuant class object containing ten pre-computed samples using the
"drews" copy number signature methodology.
}
\keyword{datasets}
