% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CNobj.mac}
\alias{CNobj.mac}
\title{CNobj.mac}
\format{
A CNQuant class object using the "mac" methodology
}
\usage{
data(CNobj.mac)
}
\description{
Example CNQuant class object containing ten pre-computed samples using the
"mac" copy number signature methodology.
}
\keyword{datasets}
