% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CNQuant-class}
\alias{CNQuant-class}
\alias{CNQuant}
\title{CNQuant object}
\description{
CNQuant object
}
\section{Slots}{

\describe{
\item{\code{segments}}{list}

\item{\code{featData}}{list}

\item{\code{featFitting}}{list}

\item{\code{samplefeatData}}{data.frame}

\item{\code{ExpData}}{ExpQuant}
}}

