% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCNQuant.R
\name{createCNQuant}
\alias{createCNQuant}
\title{createCNQuant}
\usage{
createCNQuant(
  data = NULL,
  experimentName = "defaultExperiment",
  build = "hg19"
)
}
\arguments{
\item{data}{Unrounded absolute copy number data}

\item{experimentName}{A name for the experiment (default: defaultExperiment)}

\item{build}{A genome build specified as either hg19 or hg38 (default: hg19)}
}
\value{
A CNQuant class object
}
\description{
createCNQuant
}
