% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drews2022_CX3CX2_Clinical_classifier}
\alias{Drews2022_CX3CX2_Clinical_classifier}
\title{Drews2022_CX3CX2_Clinical_classifier}
\format{
A list of 2 numeric vectors of length 2
}
\usage{
data(Drews2022_CX3CX2_Clinical_classifier)
}
\description{
List of mean and scaling factors for signatures 2 and 3 for the prediction of
platinum status used in Drews 2022
}
\keyword{datasets}
