% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drews2022_TCGA_Mixture_Models}
\alias{Drews2022_TCGA_Mixture_Models}
\title{Drews2022_TCGA_Mixture_Models}
\format{
A list of 5 data.frames
}
\usage{
data(Drews2022_TCGA_Mixture_Models)
}
\description{
List of mixture model mean, standard deviation, and weight for each copy number
feature used in Drews 2022
}
\keyword{datasets}
