% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drews2022_TCGA_Scaling_Variables}
\alias{Drews2022_TCGA_Scaling_Variables}
\title{Drews2022_TCGA_Scaling_Variables}
\format{
A list of 2 numeric vectors of length 17
}
\usage{
data(Drews2022_TCGA_Scaling_Variables)
}
\description{
List of mean and scaling factors for signatures used in Drews 2022
}
\keyword{datasets}
