% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drews2022_TCGA_Signature_Thresholds}
\alias{Drews2022_TCGA_Signature_Thresholds}
\title{Drews2022_TCGA_Signature_Thresholds}
\format{
A numeric vector of length 17
}
\usage{
data(Drews2022_TCGA_Signature_Thresholds)
}
\description{
Numeric vector containing signature-specific thresholds used in activity
calculations used in Drews 2022
}
\keyword{datasets}
