% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drews2022_TCGA_Signatures}
\alias{Drews2022_TCGA_Signatures}
\title{Drews2022_TCGA_Signatures}
\format{
A 17 by 43 numeric matrix
}
\usage{
data(Drews2022_TCGA_Signatures)
}
\description{
Signature-by-component matrix for 17 derived signatures as used in Drews 2022
}
\keyword{datasets}
