% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DrewsMacintyreMapping}
\alias{DrewsMacintyreMapping}
\title{DrewsMacintyreMapping}
\format{
A list of length 6 containing data.frame and matrix type data
}
\usage{
data(DrewsMacintyreMapping)
}
\description{
List of tables and matrices quantifying the similarity between
Drews et al. 2022 and Mactintyre et al. 2018 CIN signature definitions and
activities using computed cosine similarity across 478 TCGA gold standard samples.
}
\details{
\itemize{
\item map_drews_definitions: data.frame containing the maximum cosine similarity for Drews et al. 2022 signature definitions to each of the Macintyre et al. 2018 signature definitions.
\item map_mac_definitions: data.frame containing the maximum cosine similarity for Macintyre et al. 2018 signature definitions to each of the Drews et al. 2022 signature definitions.
\item cosine_definitions: matrix containing all-vs-all cosine similarity between Drews et al. 2022 and Macintyre et al. 2018 signature definitions.
\item map_drews_activity: data.frame containing the maximum cosine similarity for Drews et al. 2022 signature activities to each of the Macintyre et al. 2018 signature activities.
\item map_mac_activity: data.frame containing the maximum cosine similarity for Macintyre et al. 2018 signature activities to each of the Drews et al. 2022 signature activities
\item cosine_activity: matrix containing all-vs-all cosine similarity between Drews et al. 2022 and Macintyre et al. 2018 signature activities.
}
}
\keyword{datasets}
