% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateActivityThresholds.R
\name{estimateThresholds}
\alias{estimateThresholds}
\title{estimateThresholds}
\usage{
estimateThresholds(
  object = NULL,
  iters = 1000,
  SDPROP = 20,
  FINALWIDTH = 0.1,
  RANGECNAS = 0.1,
  method = "drewsV2",
  minmaxNorm = FALSE,
  orderOri = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{object}{An object of class SigQuant or PLACEHOLDERCLASS containing computed features, signature definitions, siganture activities, and feature mixture models.}

\item{iters}{Number of noise simulation iterations to perform to estimate thresholds.}

\item{method}{Feature and mixture model method to use for computing simulated features and signature activities}

\item{parallel}{Use doFuture parallelisation to run iterations set by future::plan()}
}
\value{
data.frame containing series of signature-specific thresholds (Default used = Thresh_ZeroGMM_0.05)
}
\description{
estimateThresholds
}
