% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCIN.R
\name{exportCIN}
\alias{exportCIN}
\title{exportCIN}
\usage{
exportCIN(
  object,
  outputDir = NULL,
  outputPrefix = NULL,
  includeExpName = FALSE,
  sep = "\\t",
  fullExport = FALSE
)
}
\arguments{
\item{object}{CNQuant or SigQuant object}

\item{outputDir}{Output location provided as a writable file directory. This
defaults to the current working directory.}

\item{outputPrefix}{A prefix added to the beginning of exported files.}

\item{includeExpName}{Include the object experiment name in the file name
(default: FALSE).}

\item{sep}{Default file separator used in writing files (Default: '\\t')}

\item{fullExport}{Provide a full export of all data contained within the
provided object, including reference data, feature values, models, and
supporting information (default: FALSE).}
}
\description{
This function exports data from a \code{CNQuant} or \code{SigQuant} class object.
}
