% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmixturepois.R
\name{fitMixturePois}
\alias{fitMixturePois}
\title{fitMixturePois}
\usage{
fitMixturePois(
  component = NULL,
  seed = 77777,
  model_selection = "BIC",
  min_prior = 0.001,
  niter = 1000,
  nrep = 1,
  min_comp = 2,
  max_comp = 10
)
}
\arguments{
\item{component}{two-dimensional data.frame containing sample names and values of computed copy number features generated by calculateFeatures()}

\item{seed}{Numeric or string as a input value for set.seed() function - Currently disabled (Default: '77777')}

\item{model_selection}{String to specify model selection method of available "AIC", "BIC", or "ICL" (default: "BIC")}

\item{min_prior}{FLXcontrol value - Minimum prior probability of clusters, components falling below this threshold are removed during the iteration (default: 0.001)}

\item{niter}{FLXcontrol value - Maximum number of iterations (default: 1000)}

\item{nrep}{Number of random initialisations to run per mixture component step (default: 1)}

\item{min_comp}{Minimum of mixture components to compute (default: 2)}

\item{max_comp}{Minimum of mixture components to compute (default: 10)}
}
\value{
data.table containing the mean and weights of fitted poisson mixture
models
}
\description{
Fit mixture model using flexmix to set of components with Poisson
distributions
}
