% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gap_hg19}
\alias{gap_hg19}
\title{gap_hg19}
\format{
A data frame with 457 rows and 9 variables
}
\usage{
data(gap_hg19)
}
\description{
Chromosomal banding and position of genomic features in genome build hg19
}
\keyword{datasets}
