% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gap_hg38}
\alias{gap_hg38}
\title{gap_hg38}
\format{
A data frame with 827 rows and 9 variables
}
\usage{
data(gap_hg38)
}
\description{
Chromosomal banding and position of genomic features in genome build hg38
}
\keyword{datasets}
