% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getActivities.R
\docType{methods}
\name{getActivities}
\alias{getActivities}
\alias{getActivities,SigQuant-method}
\title{getActivities}
\usage{
getActivities(object, type = "threshold")

\S4method{getActivities}{SigQuant}(object, type = "threshold")
}
\arguments{
\item{object}{SigQuant object}

\item{type}{type of copy number signature matrix to return 'raw', 'norm', 'threshold', or 'scaled'}
}
\value{
A numeric matrix
}
\description{
Extracts signature activities from a SigQuant object.
}
