% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getDefinitions.R
\docType{methods}
\name{getDefinitions}
\alias{getDefinitions}
\alias{getDefinitions,SigQuant-method}
\title{getDefinitions}
\usage{
getDefinitions(object, normalise = FALSE)

\S4method{getDefinitions}{SigQuant}(object, normalise = FALSE)
}
\arguments{
\item{object}{SigQuant object}

\item{normalise}{Logical indicating whether to normalise signatures
definitions by feature and signature (default: FALSE)}
}
\value{
matrix containing the signature definitions
}
\description{
getDefinitions
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  t478 <- TCGA_478_Samples_SNP6_GOLD
  subsample <- unique(t478$sample)[1:10]
  t478 <- t478[t478$sample \%in\% subsample]
  cnobj <- quantifyCNSignatures(t478)
  getDefinitions(cnobj)

}
\seealso{
\code{\link[=plotDefinitions]{plotDefinitions()}}
}
