% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getExperiment.R
\docType{methods}
\name{getExperiment}
\alias{getExperiment}
\alias{getExperiment,CNQuant-method}
\title{getExperiment}
\usage{
getExperiment(object)

\S4method{getExperiment}{CNQuant}(object)
}
\arguments{
\item{object}{CNQuant object}
}
\value{
A ExpQuant class object
}
\description{
Extracts and returns copy number features from copy number profiles in a CNQuant object.
}
