% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getFeatures.R
\docType{methods}
\name{getFeatures}
\alias{getFeatures}
\alias{getFeatures,CNQuant-method}
\title{getFeatures}
\usage{
getFeatures(object, feat = NULL)

\S4method{getFeatures}{CNQuant}(object, feat = NULL)
}
\arguments{
\item{object}{CNQuant object}

\item{feat}{name or vector names of features to return. By default returns all features.}
}
\value{
list containing the extracted feature data
}
\description{
getFeatures
}
