% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSamplefeatures.R
\docType{methods}
\name{getSamplefeatures}
\alias{getSamplefeatures}
\alias{getSamplefeatures,CNQuant-method}
\title{getSamplefeatures}
\usage{
getSamplefeatures(object)

\S4method{getSamplefeatures}{CNQuant}(object)
}
\arguments{
\item{object}{CNQuant object}
}
\value{
A data.frame containing the sample-specific features (ploidy, segment
count, etc.)
}
\description{
Extracts sample feature data from a \code{CNQuant} class object.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- createCNQuant(TCGA_478_Samples_SNP6_GOLD)
  SampleFeats <- getSamplefeatures(cnobj[1:10])
}
\seealso{
\code{\link[=addsampleFeatures]{addsampleFeatures()}}
}
