% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSamples.R
\docType{methods}
\name{getSamples}
\alias{getSamples}
\alias{getSamples,CNQuant-method}
\title{getSamples}
\usage{
getSamples(object)

\S4method{getSamples}{CNQuant}(object)
}
\arguments{
\item{object}{CNQuant object}
}
\value{
character vector containing all available sample names
}
\description{
Returns sample names from a CNQuant class object.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- createCNQuant(TCGA_478_Samples_SNP6_GOLD)
  Samples <- getSamples(cnobj[1:10])
}
