% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSegments.R
\docType{methods}
\name{getSegments}
\alias{getSegments}
\alias{getSegments,CNQuant-method}
\title{getSegments}
\usage{
getSegments(object)

\S4method{getSegments}{CNQuant}(object)
}
\arguments{
\item{object}{CNQuant object}
}
\value{
A data.frame
}
\description{
Extracts copy number segment data from a CNQuant object.
}
