% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSignatures.R
\docType{methods}
\name{getSignatures}
\alias{getSignatures}
\alias{getSignatures,SigQuant-method}
\title{getSignatures}
\usage{
getSignatures(object, type = "threshold")

\S4method{getSignatures}{SigQuant}(object, type = "threshold")
}
\arguments{
\item{object}{SigQuant object}

\item{type}{type of copy number signature matrix to return 'raw', 'norm', 'threshold', or 'scaled'}
}
\value{
A numeric matrix
}
\description{
Extracts signatures names from a SigQuant object.
}
