% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19.chrom.sizes}
\alias{hg19.chrom.sizes}
\title{hg19.chrom.sizes}
\format{
A data frame with 24 rows and 2 variables
}
\usage{
data(hg19.chrom.sizes)
}
\description{
Chromosomal lengths for genome build hg19
}
\keyword{datasets}
