% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Macintyre2018_OV_Mixture_Models}
\alias{Macintyre2018_OV_Mixture_Models}
\title{Macintyre2018_OV_Mixture_Models}
\format{
A list of 6 data.frames
}
\usage{
data(Macintyre2018_OV_Mixture_Models)
}
\description{
List of mixture model mean, standard deviation, and weight for each copy number
feature used in Macintyre 2018
}
\keyword{datasets}
