% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Macintyre2018_OV_Signatures}
\alias{Macintyre2018_OV_Signatures}
\title{Macintyre2018_OV_Signatures}
\format{
A 7 by 36 numeric matrix
}
\usage{
data(Macintyre2018_OV_Signatures)
}
\description{
Signature-by-component matrix for 7 derived signatures as used in Macintyre 2018
}
\keyword{datasets}
