% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Macintyre2018_OV_Signatures_normalised}
\alias{Macintyre2018_OV_Signatures_normalised}
\title{Macintyre2018_OV_Signatures_normalised}
\format{
A 7 by 36 numeric matrix
}
\usage{
data(Macintyre2018_OV_Signatures_normalised)
}
\description{
Signature-by-component matrix for 7 derived signatures as used in Macintyre 2018
where column sums are normalised to 1.
}
\keyword{datasets}
