% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Macintyre2018_OV_Signatures_t}
\alias{Macintyre2018_OV_Signatures_t}
\title{Macintyre2018_OV_Signatures_t}
\format{
A 7 by 36 numeric matrix
}
\usage{
data(Macintyre2018_OV_Signatures_t)
}
\description{
Transposed Signature-by-component matrix for 7 derived signatures as used in Macintyre 2018
TEMPORARY DATA
}
\keyword{datasets}
