% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegments.R
\name{plotSegments}
\alias{plotSegments}
\title{plotSegments}
\usage{
plotSegments(object = NULL, sample = NULL, cn.max = 15)
}
\arguments{
\item{object}{A CNQuant or SigQuant class object}

\item{sample}{A vector of length 1 containing either a sample name or sample index}

\item{cn.max}{Maximum copy number to plot - Values over this are truncated to fit}
}
\value{
plot
}
\description{
Plot the segment data for a given sample stored in a CNQuant or SigQuant class object
}
