% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectrum.R
\name{plotSpectrum}
\alias{plotSpectrum}
\title{plotSpectrum}
\usage{
plotSpectrum(object, sample = NULL, cols = NULL)
}
\arguments{
\item{object}{A CNQuant class object containing}

\item{sample}{vector of length 1 containing either a sample name or sample
index}

\item{cols}{Vector of colours of the same length as number of copy number
components.}
}
\value{
plot
}
\description{
Plot the copy number feature components for a given sample in a CNQuant class object
containing computed sample by component. Default ordering by
decreasing exposure to signature CX1.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  t478 <- TCGA_478_Samples_SNP6_GOLD
  subsample <- unique(t478$sample)[1:10]
  t478 <- t478[t478$sample \%in\% subsample]
  cnobj <- quantifyCNSignatures(t478)
  plotSpectrum(cnobj,sample = 1)
}
\seealso{
\code{\link[=getSampleByComponent]{getSampleByComponent()}}
}
