% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SIGobj.drews}
\alias{SIGobj.drews}
\title{SIGobj.drews}
\format{
A CNQuant class object using the "drews" methodology
}
\usage{
data(SIGobj.drews)
}
\description{
Example SigQuant class object containing ten pre-computed samples using the
"drews" copy number signature methodology.
}
\keyword{datasets}
