% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SIGobj.mac}
\alias{SIGobj.mac}
\title{SIGobj.mac}
\format{
A CNQuant class object using the "mac" methodology
}
\usage{
data(SIGobj.mac)
}
\description{
Example SigQuant class object containing ten pre-computed samples using the
"mac" copy number signature methodology.
}
\keyword{datasets}
