% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SigQuant-class}
\alias{SigQuant-class}
\alias{SigQuant}
\title{SigQuant}
\description{
SigQuant
}
\section{Slots}{

\describe{
\item{\code{activities}}{list}

\item{\code{signature.model}}{character}

\item{\code{backup.signatures}}{matrix}

\item{\code{backup.thresholds}}{numeric}

\item{\code{backup.scale}}{list}

\item{\code{backup.scale.model}}{character}
}}

