% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SigQuant.R
\name{[,SigQuant,numeric,missing-method}
\alias{[,SigQuant,numeric,missing-method}
\title{Extract}
\usage{
\S4method{[}{SigQuant,numeric,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{object to extract subset}

\item{i}{Indices of subset}

\item{j}{not used}

\item{...}{not used}

\item{drop}{not used}
}
\description{
Extract
}
