% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_478_Samples_SNP6_GOLD}
\alias{TCGA_478_Samples_SNP6_GOLD}
\title{TCGA_478_Samples_SNP6_GOLD}
\format{
A data.frame with 5 columns and 65098 rows.
}
\usage{
data(TCGA_478_Samples_SNP6_GOLD)
}
\description{
Example data set of 478 pan-cancer samples with detectable CIN in the correct
input format for generating copy number signatures. Data was generated during
the development and submission of "A pan-cancer compendium of CIN" drews et
al. 2022.
}
\keyword{datasets}
