% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.sample.features}
\alias{test.sample.features}
\title{test.sample.features}
\format{
A data.frame with 4 columns and 478 rows.
}
\usage{
data(test.sample.features)
}
\description{
Example data set of additional example sample features to add to the existing
sample feature data of the 478 pan-cancer samples with detectable CIN. Data
in this dataset are dummy variables only for function testing and examples
and have no meaning.
}
\keyword{datasets}
