% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calculateActivity.R
\docType{methods}
\name{calculateActivity}
\alias{calculateActivity}
\alias{calculateActivity,CNQuant-method}
\title{calculateActivity}
\usage{
calculateActivity(object, method = "drews")

\S4method{calculateActivity}{CNQuant}(object, method = NULL)
}
\arguments{
\item{object}{CNQuant object}

\item{method}{Determines the mixture components used to calculate sum-of-posterior probabilities. Default is "drews".}
}
\value{
A SigQuant class object with four activity matrices stored in the "activities" slot
}
\description{
Calculates and returns signature activities in a SigQuant object. Works best after function calculateSampleByComponentMatrix call. \cr \cr
The output of this function is a list of four matrices, the raw signature activities, the normalised activities, the normalised and
thresholded signature activities and the normalised, thresholded and scaled activities with the scaling factors obtained from the TCGA cohort.
}
