% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/clinPredictionPlatinum.R
\docType{methods}
\name{clinPredictionPlatinum}
\alias{clinPredictionPlatinum}
\alias{clinPredictionPlatinum,SigQuant-method}
\title{clinPredictionPlatinum}
\usage{
clinPredictionPlatinum(object)

\S4method{clinPredictionPlatinum}{SigQuant}(object)
}
\arguments{
\item{object}{SigQuant object}
}
\value{
A vector with "Predicted sensitive" or "Predicted resistant" for all
samples in the input object.
}
\description{
The function takes signature activities based on Drews et al. 2022
methodology and predicts response to platinum-based chemo-therapeutics.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- createCNQuant(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- calculateFeatures(cnobj[1:10],method="drews")
  cnobj <- calculateSampleByComponentMatrix(cnobj)
  cnobj <- calculateActivity(cnobj)
  cnobj <- clinPredictionPlatinum(cnobj)
}
