% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmixturepois.R
\name{fitMixturePois}
\alias{fitMixturePois}
\title{fitMixturePois}
\usage{
fitMixturePois(
  component = NULL,
  seed = NULL,
  model_selection = "BIC",
  min_prior = 0.001,
  niter = 1000,
  nrep = 1,
  min_comp = 2,
  max_comp = 10,
  iters = 1,
  parallel = FALSE
)
}
\arguments{
\item{component}{two-dimensional data.frame containing sample names and
values of computed copy number features generated by calculateFeatures()}

\item{seed}{Numeric or string as a input value for set.seed() function -
Currently disabled (Default: NULL)}

\item{model_selection}{String to specify model selection method of available
"AIC", "BIC", or "ICL" (default: "BIC")}

\item{min_prior}{FLXcontrol value - Minimum prior probability of clusters,
components falling below this threshold are removed during the iteration
(default: 0.001)}

\item{niter}{FLXcontrol value - Maximum number of iterations per component k
(default: 1000)}

\item{nrep}{Number of random initialisations to run per mixture component
step (default: 1)}

\item{min_comp}{Minimum of mixture components k to compute (default: 2)}

\item{max_comp}{Minimum of mixture components k to compute (default: 10)}

\item{iters}{Number of repeated runs to perform of Poisson mixture model
fitting. The modal value for number of components is selected from the
total number of iterations and the model with the best model selection
value returned (Default: 1)}

\item{parallel}{Use doFuture parallelism (default: FALSE)}
}
\value{
data.table containing the mean and weights of fitted poisson mixture
models
}
\description{
Fit mixture model using flexmix to set of components with Poisson
distributions
}
