% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getExperiment.R
\docType{methods}
\name{getExperiment}
\alias{getExperiment}
\alias{getExperiment,CNQuant-method}
\title{getExperiment}
\usage{
getExperiment(object)

\S4method{getExperiment}{CNQuant}(object)
}
\arguments{
\item{object}{CNQuant object}
}
\value{
A \code{ExpQuant} class object
}
\description{
Returns the experiment details from a \code{CNQuant} class object.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- createCNQuant(TCGA_478_Samples_SNP6_GOLD)
  getExperiment(cnobj)
}
\seealso{
\linkS4class{ExpQuant}
}
