% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getFeatures.R
\docType{methods}
\name{getFeatures}
\alias{getFeatures}
\alias{getFeatures,CNQuant-method}
\title{getFeatures}
\usage{
getFeatures(object, feat = NULL)

\S4method{getFeatures}{CNQuant}(object, feat = NULL)
}
\arguments{
\item{object}{CNQuant object containing copy number features}

\item{feat}{name or vector names of features to return. By default returns
all features.}
}
\value{
list containing the extracted feature data
}
\description{
Returns the copy number features from a \code{CNQuant} class object. Features
available will be those extracted as defined by the copy number signature
method used.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- createCNQuant(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- calculateFeatures(cnobj[1:10],method="drews")
  features <- getFeatures(cnobj)
}
