% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSamplefeatures.R
\docType{methods}
\name{getSamplefeatures}
\alias{getSamplefeatures}
\alias{getSamplefeatures,CNQuant-method}
\title{getSamplefeatures}
\usage{
getSamplefeatures(object)

\S4method{getSamplefeatures}{CNQuant}(object)
}
\arguments{
\item{object}{CNQuant object}
}
\value{
A data.frame
}
\description{
Extracts sample feature data from a CNQuant object.
}
