% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSamples.R
\docType{methods}
\name{getSamples}
\alias{getSamples}
\alias{getSamples,CNQuant-method}
\title{getSamples}
\usage{
getSamples(object)

\S4method{getSamples}{CNQuant}(object)
}
\arguments{
\item{object}{CNQuant object}
}
\value{
A character vector
}
\description{
Extracts sample names from a CNQuant object.
}
