% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSegments.R
\docType{methods}
\name{getSegments}
\alias{getSegments}
\alias{getSegments,CNQuant-method}
\title{getSegments}
\usage{
getSegments(object)

\S4method{getSegments}{CNQuant}(object)
}
\arguments{
\item{object}{CNQuant object}
}
\value{
A data.frame containing the segment table for all samples.
}
\description{
Extracts copy number segment data from a \code{CNQuant} class object.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- createCNQuant(TCGA_478_Samples_SNP6_GOLD)
  segments <- getSegments(cnobj[1:10])
}
