% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDefinitions.R
\name{plotDefinitions}
\alias{plotDefinitions}
\title{plotDefinitions}
\usage{
plotDefinitions(object, cols = NULL, plot.dim = NULL)
}
\arguments{
\item{object}{A SigQuant class object containing}

\item{cols}{Vector of colours of the same length as number of copy number
components.}

\item{plot.dim}{vector length two giving the plotting grid dimensions}
}
\value{
plot
}
\description{
Plot the copy number feature components for a given sample in a SigQuant class object
containing computed sample by component. Default ordering by
decreasing exposure to signature CX1.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  t478 <- TCGA_478_Samples_SNP6_GOLD
  subsample <- unique(t478$sample)[1:10]
  t478 <- t478[t478$sample \%in\% subsample]
  cnobj <- quantifyCNSignatures(t478)
  plotDefinitions(cnobj)
}
\seealso{
\code{\link[=getDefinitions]{getDefinitions()}}
}
