% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleByComponent.R
\name{plotSampleByComponent}
\alias{plotSampleByComponent}
\title{plotSampleByComponent}
\usage{
plotSampleByComponent(object = NULL, ...)
}
\arguments{
\item{object}{a CNQuant or SigQuant class object}

\item{...}{additional parameters passed to \link[stats]{heatmap}}
}
\value{
plot
}
\description{
Plots a heatmap of the sample-by-component matrix
}
