% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegments.R
\name{plotSegments}
\alias{plotSegments}
\title{plotSegments}
\usage{
plotSegments(object = NULL, sample = NULL, cn.max = 15)
}
\arguments{
\item{object}{CNQuant or SigQuant class object}

\item{sample}{vector of length 1 containing either a sample name or sample
index}

\item{cn.max}{maximum copy number to plot - Values greater than this are
truncated to the specified value}
}
\value{
plot
}
\description{
Plot the segment data for a given sample stored in a CNQuant or SigQuant
class object.
}
\examples{
  data(TCGA_478_Samples_SNP6_GOLD)
  cnobj <- createCNQuant(TCGA_478_Samples_SNP6_GOLD)
  segments <- plotSegments(cnobj,sample=1,cn.max=15)
}
\seealso{
\code{\link[=getSegments]{getSegments()}}
}
