% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/quantifyCNSignatures.R
\docType{methods}
\name{quantifyCNSignatures}
\alias{quantifyCNSignatures}
\alias{quantifyCNSignatures,ANY-method}
\title{quantifyCNSignatures}
\usage{
quantifyCNSignatures(
  object,
  experimentName = "Default",
  method = "drews",
  cores = 1,
  build = "hg19"
)

\S4method{quantifyCNSignatures}{ANY}(
  object,
  experimentName = "Default",
  method = "drews",
  cores = 1,
  build = "hg19"
)
}
\arguments{
\item{object}{input data (segment file - file path or data.frame)}

\item{experimentName}{A user-specified name of the experiment}

\item{method}{The method used for calculating the signature activities. Default is "drews"}

\item{cores}{Number of threads/cores to use for parallel processing}

\item{build}{Genome build to use (default: hg19)}
}
\value{
A SigQuant class object with four activity matrices stored in the "activities" slot
}
\description{
This function takes a copy number profile as input and returns signature activities.
}
